gestion d'accessibilité des pages en fonction de l'utilisateur et affichage de l'auteur

+++ delete-article.php
@@ -1,17 +1,34 @@
 <?php 
-/**
- * @var ArticleDAO
- */
 
-    $articleDAO = require_once'./database/models/ArticleDAO.php';
+    require_once __DIR__.'/database/database.php';
+    require_once __DIR__.'/database/security.php';
+
+    $currentUser = isLoggedIn();
+
+    if(!$currentUser) {
+        header('Location:/');
+    }
+
+
+    /**
+     * @var ArticleDAO
+     */
+
+        $articleDAO = require_once'./database/models/ArticleDAO.php';
 
 
 
-$_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
-$id = $_GET['id'] ?? '';
+    $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
+    $id = $_GET['id'] ?? '';
 
 
-if($id) {
-    $articleDAO->deleteOne($id);
-}
-header('Location: /');
+    if($id ) {
+        $article = $articleDAO->getOne($id);
+
+        if($currentUser['id'] === $article['author']){
+            $articleDAO->deleteOne($id);
+        }else {
+            header('Location: /');
+        }
+    }
+    header('Location: /');
